Add files via upload
The Python code using CAPM and Monte Carlo simulation is a financial tool used to allocate portfolio weights. It generates 50,000 portfolios and calculates their returns and risk based on the Efficient Frontier. The Efficient Frontier is a set of optimal portfolios that offer the highest expected return for a given level of risk or the lowest risk for a given level of expected return.

The code uses CAPM(Capital Asset Pricing Model) to calculate the expected return on stock daily annualized and daily alpha for stocks. The CAPM estimates the expected return on an asset based on its risk and the expected return of the overall market. The Monte Carlo simulation generates random values for the stock's return and risk, which allows the model to simulate different scenarios and calculate the probability of different outcomes.

Although CAPM is a widely used model in finance, it has some limitations. One limitation is that it assumes that all investors have the same beliefs about expected returns, risk, and correlation of assets. This is known as the assumption of homogeneous expectations, which may not always hold in the real world. Another limitation is that CAPM assumes that the relationship between expected return and beta is linear, which may not be true.

CAPM also relies heavily on historical data, which may not always accurately predict future returns and risk. Additionally, CAPM assumes that the market is efficient and that all information is readily available to investors, which may not always be the case. Finally, the model does not account for other factors that may influence asset returns, such as changes in interest rates, geopolitical events, and macroeconomic factors. As such, investors should use CAPM in conjunction with other models and tools to make informed investment decisions.